+++ patches/scripts/199-patch_cockpit_unsecure.sh
@@ -1,5 +1,0 @@
-[ "$FREETZ_PATCH_COCKPIT_UNSECURE" == "y" ] || return 0
-echo1 "applying cockpit unsecure settings patch"
-
-cockpit_pruner "osInformations"
-